More work on save files. Should be entirely working now.
If you don't want a variable being written to the save file at all then
you should make it a transient variable. This will stop it from being
serialized and deserialized into and from JSON. You should make things
such as Sprites transient as these should be created automatically when
an Entity or something else is instantiated.

Rather than saving the World name in the save data I may change it to
save the World object instead. This would be best for storing all the
entities that aren't the player.